django/requirements.txt
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==2.5.2
django-crispy-forms==1.7.2
django-fontawesome==1.0
django-nested-admin==3.1.3
psycopg2-binary==2.7.7
pygal==2.4.0
python-monkey-business==1.0.0
pytz==2019.1
PyYAML==5.1
six==1.12.0
